Change naming of imported macros
Previously, macros imported from external files named the macros
according to their name in the file and appended a % character

This could cause name clashes if more than one file is imported
that uses the same macro name(s).

This commit changes the naming to prepend the macro name assigned
when creating the import macro to the macro name from the file. The
two names are separated by a dot and the trailing % sign is no
longer used. E.g. foo.bar where foo is the name given to the macro
in the macro editor and bar is the macro name read from the imported
file.