Modify macro name validation in FmMacroEd
Change of macro name validation in TVInfo.IsValidMacroName to disallow
empty macro names broke validation in FmMacroEd's name control: it
became impossible to delete a macro name from the box.

The method of validation was changed to permit deletion of the edit box
contents while still refusing to let a macro with an empty name from
being included in the macro list.