Remove escaping from double quotes in Version.vi
Since bug at issue #57 was fixed any quotes in a string table entry in
the .vi file are automatically escaped when writing the .rc file.
Version.vi contained ready-escaped double quotes as a work around for
the bug. These are now being double-escaped following the bug fix, so
the \ escape characters were removed from the quotes in Version.vi