Fix bug in UUtils.RemoveExtension function
This was the cause of bug #62 that affected the <SHORTFNAME> field.

This routine was stripping anything after the first occurence of the
dot extension delimiter. Tmeant that file names containing multiple dots
were getting chopped in the wrong place. E.g. `foo.bar.txt` was being
truncated to `foo` when it should have been truncated to `foo.bar`.

Fixes #62