Refactor UVInfo unit and TVIInfo class
All macro code was pulled out into a new macro class, TMacros, in the
new UMacros unit.

New TVIFile class in the new UVIFile unit was created to encapsulate the
.vi file. This got round the problem that TMacros needed access to
information about the file that was previously maintained in UVInfo,
causing a possible circular unit reference.

Updated forms affected by the change of the interface to macros.

This was a major refactoring.