Replace view macros dlg with a new custom dlg
Added new FmViewMacros unit to the project. It contains the new dialogue
box. This is a tabbed dialogue box that displays macro details in a list
view in one tab and warnings in a memo control in another tab. The
warning tab is hidden if there are no warnings.

Revised FmMain to use the new dialog box.