Revise how macros are handled in TVerNumEditor
TVerNumEditor was changed to have valid macros added one at a time by
the calling code rather than passing a string list of names to it.

Renamed some components of TVerNumEditor to clarify that they refer to
macros.

Changed so that the macro insertion button is now disabled as well as
the macro combo box when there are no macros available.

Modified calling code in TMainForm to add macros one at a time to
TVerNumEditor. It does this by enumerating the array or resolved macros
returned by TMacros.GetResolvedMacros insead of calling
TMacros.ListResolvedMacroNames.